Lift MockDB type to its own file.
